moved search suggestions list item to separate file
