+++ my-app/src/components/dumb/SearchSuggestionsList/SearchSuggestionsItem.tsx
@@ -1,0 +1,19 @@
+import { Link } from 'react-router-dom';
+import { Movie } from '../../../api/types';
+
+function SearchSuggestionsItem(
+  {id, title, releaseDate}: Movie
+) {
+  const path = `/movie/${id}`;
+  const year = releaseDate.slice(0, 4);
+
+  return (
+    <li className='search-list__item'>
+      <Link to={path}>
+        {`${title} (${year})`}
+      </Link>
+    </li>
+  );
+}
+
+export default SearchSuggestionsItem;
